Two patches make docker-snap work in snappy world.
1.Currently docker only checks the normal user info from /etc/passwd
and /etc/group folder. And the extrauser/group info is stored in
/var/lib/extrausers/passwd,/var/lib/extrausers/group folder. We add a
support to let docker lookup extrausers info if uid/gid can not be found
at system path.
2.New "chroot" policy results in new container root fallbacks into "/"
folder.So when running a container, open it under the "/" root folder
which causes the apparmor deines error. We use the realChroot function
in this case just like it runs in 1.11.2.