Build golang1.7 from source.
As of now, go plugin doesn't support to version option and
we're not able to specify go 1.7 as the preferred compiler to
build each docker and its components.
An open bug can be found here
https://bugs.launchpad.net/snapcraft/+bug/1616985

Meanwhile, golang1.7 deb package for the archs[armhf,arm64,ppc]
can not be found in the ppa we imported.
https://launchpad.net/~jonathonf/+archive/ubuntu/golang-1.7

So in order to release docker-snap for all archs, we have to
create custom plugin and build golang-1.7 from source as
the compiler that we're going to use to build docker snap.