many: support a daemon.json config file for dockerd
This allows us to easily specify and change config values for the dockerd daemon without recompiling or changing the snap. This config file is installed from $SNAP on install with a hook and lives in $SNAP_COMMON so it persists between refreshes.
This is also necessary so we can switch to the aufs storage-driver, as the default overlayfs shouldn't be used in snaps and could break in unexpected ways in the future.