Update dockerd-wrapper
Better variant of fixing stale files preventing docker starting after an unclean shutdown. Move all `/run` files that dockerd expects to be in `/run`  and thus cleaned up by a reboot to `/run/snap.$SNAP_INSTANCE_NAME`. This is a snap-specific location that is guaranteed to be accessible and also cleaned up by virtue of actually being stored in `/run` on the host system.