add: utility script to upgrade docker-snap version
- It works by running it from the base root of the project:
```
./.github/workflows/upgrade.sh
```
It generates the necessary changes on snap/snapcraft.yaml
by fetching the latest version available for Moby
using the GitHub api.
And then based on the Dockerfile for the latest version, it
fetches the necessary tags for the other parts of docker-snap.
It only changes the needed code and not the comments.

Signed-off-by: Lincoln Wallace <lincoln.wallace@canonical.com>